Calculate UN WPP-Ajusted population density covariates covariates

Extract population density data at point locations using SpatRaster object from process_sedac_population. Function returns a data frame containing population values at user-defined sites. Unique column names reflect variable name and circular buffer.

Usage

calc_sedac_population(from, locs, locs_id = NULL, radius = 0, fun = "mean")

Arguments

from

SpatRaster(1). Cleaned SpatRaster object that has been returned from process_sedac_population containing GEOS-CF variable data.

locs

data.frame, characater to file path, SpatVector, or sf object.

locs_id

character(1). Column within locations CSV file containing identifier for each unique coordinate location.

radius

integer(1). Circular buffer distance around site locations. (Default = 0).

fun

character(1). Function used to summarize multiple raster cells within sites location buffer (Default = mean).

Value

a data.frame object;

Author

Mitchell Manware