Calculate Sum of Exponentially Decaying Contributions (SEDC) covariates

Calculate Sum of Exponentially Decaying Contributions (SEDC) covariates

Usage

calc_sedc(
  from = NULL,
  locs = NULL,
  locs_id = NULL,
  sedc_bandwidth = NULL,
  target_fields = NULL,
  geom = FALSE
)

Arguments

from

SpatVector object. Locations where each SEDC is calculated.

locs

SpatVector object. Locations where the sum of SEDCs are calculated.

locs_id

character(1). Name of the unique id field in point_to.

sedc_bandwidth

numeric(1). Distance at which the source concentration is reduced to exp(-3) (approximately -95 %)

target_fields

character(varying). Field names in characters.

geom

logical(1). Should the function return a SpatVector? Default is FALSE. The coordinate reference system of the SpatVector is that of from.

Value

a data.frame (tibble) or SpatVector object with input field names with a suffix "_sedc" where the sums of EDC are stored. Additional attributes are attached for the EDC information.

`attr(result, "sedc_bandwidth")“: the bandwidth where concentration reduces to approximately five percent

`attr(result, "sedc_threshold")“: the threshold distance at which emission source points are excluded beyond that

Note

The function is originally from chopin Distance calculation is done with terra functions internally. Thus, the function internally converts sf objects in point_* arguments to terra. The threshold should be carefully chosen by users.

References

Messier KP, Akita Y, Serre ML (2012). “Integrating Address Geocoding, Land Use Regression, and Spatiotemporal Geostatistical Estimation for Groundwater Tetrachloroethylene.” Environmental Science & Technology, 46(5), 2772–2780. ISSN 0013-936X, doi:10.1021/es203152a . Wiesner C (????). “Euclidean Sum of Exponentially Decaying Contributions Tutorial.” https://mserre.sph.unc.edu/BMElab_web/SEDCtutorial/index.html.

Author

Insang Song

Examples

library(terra)
#> terra 1.7.78
library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
set.seed(101)
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(ncpath)
nc <- terra::project(nc, "EPSG:5070")
pnt_locs <- terra::centroids(nc, inside = TRUE)
pnt_locs <- pnt_locs[, "NAME"]
pnt_from <- terra::spatSample(nc, 100L)
pnt_from$pid <- seq(1, 100)
pnt_from <- pnt_from[, "pid"]
pnt_from$val1 <- rgamma(100L, 1, 0.05)
pnt_from$val2 <- rgamma(100L, 2, 1)

vals <- c("val1", "val2")
calc_sedc(pnt_locs, pnt_from, "NAME", 1e5, vals)
#> Joining with `by = join_by(from_id)`
#> Joining with `by = join_by(to_id)`
#> Joining with `by = join_by(from_id, to_id)`
#>             NAME       val1       val2
#> 1       Alamance  66.306697  6.7116902
#> 2      Alexander  97.987170 11.6645958
#> 3      Alleghany 135.316205  8.5703397
#> 4          Anson  50.005415  5.2276410
#> 5           Ashe 143.015620  8.4157319
#> 6          Avery  76.088990  8.7578666
#> 7       Beaufort  58.759937  6.2939195
#> 8         Bertie  70.112408  7.8469177
#> 9         Bladen  79.513358  6.8157166
#> 10     Brunswick  54.292706  5.3356971
#> 11      Buncombe  57.603317  9.6687275
#> 12         Burke  91.378099 13.5219825
#> 13      Cabarrus  64.421536  7.2742864
#> 14      Caldwell 103.478357 11.6921332
#> 15        Camden  77.730226  4.5635043
#> 16      Carteret  45.586379  6.0638371
#> 17       Caswell  46.799679  4.7822705
#> 18       Catawba  77.542299 12.2965447
#> 19       Chatham  81.981947  8.1190512
#> 20      Cherokee   3.915403  0.9588335
#> 21        Chowan  64.504924  6.0186155
#> 22          Clay   9.749667  2.4724810
#> 23     Cleveland  67.458061  9.8771979
#> 24      Columbus  48.414701  5.1153279
#> 25        Craven  61.597042  4.9988722
#> 26    Cumberland  96.540752  8.2064850
#> 27     Currituck  63.140911  4.0256414
#> 28          Dare  41.540511  3.4968348
#> 29      Davidson  87.860462 12.1240335
#> 30         Davie  90.343496 13.6229310
#> 31        Duplin  68.001564  5.3855035
#> 32        Durham  79.049439  6.2339198
#> 33     Edgecombe  70.147137  6.9895227
#> 34       Forsyth  71.195554 13.3624413
#> 35      Franklin  66.157171  6.4014195
#> 36        Gaston  48.671028  7.6455348
#> 37         Gates  64.098198  5.5317312
#> 38        Graham   7.198239  1.6872974
#> 39     Granville  49.656550  4.7207362
#> 40        Greene  90.485199  5.7627158
#> 41      Guilford  57.173967  8.4095114
#> 42       Halifax  73.885565  7.7264226
#> 43       Harnett  98.976569  8.5364366
#> 44       Haywood  38.540767  6.6754283
#> 45     Henderson  53.566765  6.8870803
#> 46      Hertford  76.272095  7.1651027
#> 47          Hoke 103.981765  8.3039185
#> 48          Hyde  67.876877  6.0577500
#> 49       Iredell  80.189217 11.8598788
#> 50       Jackson  21.292017  6.0767266
#> 51      Johnston  93.203340  7.1667761
#> 52         Jones  66.025083  4.9071098
#> 53           Lee  97.560011  9.4921343
#> 54        Lenoir  77.550455  5.1551243
#> 55       Lincoln  63.939958  9.9774137
#> 56         Macon  15.059649  4.2463505
#> 57       Madison  70.241627 10.2826394
#> 58        Martin  66.537364  7.5693036
#> 59      McDowell  67.787551 10.2585459
#> 60   Mecklenburg  50.862565  7.7914928
#> 61      Mitchell  65.031353  8.5814715
#> 62    Montgomery  59.749763  6.0405602
#> 63         Moore  78.615922  7.3924314
#> 64          Nash  64.385415  6.4241311
#> 65   New Hanover  67.931926  4.6333534
#> 66   Northampton 118.738765  9.2546833
#> 67        Onslow  77.084563  5.0114996
#> 68        Orange  72.966060  6.0369847
#> 69       Pamlico  50.073396  5.3901503
#> 70    Pasquotank  74.974591  4.7458824
#> 71        Pender  79.062124  5.6213647
#> 72    Perquimans  66.431520  5.0567758
#> 73        Person  52.598860  4.1839645
#> 74          Pitt  79.072691  5.8360247
#> 75          Polk  50.487473  7.5159326
#> 76      Randolph  70.147697  7.7840601
#> 77      Richmond  61.478506  6.1899525
#> 78       Robeson  86.300294  6.6178460
#> 79    Rockingham  44.705144  8.4209078
#> 80         Rowan  76.998697 10.2729658
#> 81    Rutherford  62.626215 10.1472106
#> 82       Sampson  70.487687  6.6277097
#> 83      Scotland  76.605109  6.8632094
#> 84        Stanly  62.852360  6.4087924
#> 85        Stokes  62.166361 16.0572751
#> 86         Surry  81.764284 12.2180431
#> 87         Swain  13.035910  2.6972329
#> 88  Transylvania  30.106402  5.9152072
#> 89       Tyrrell  67.837730  5.6893818
#> 90         Union  40.905040  5.1267190
#> 91         Vance  47.201872  5.9009680
#> 92          Wake 102.424634  7.3683230
#> 93        Warren  41.919319  6.5553331
#> 94    Washington  59.440806  5.7795964
#> 95       Watauga 108.967001 10.6908618
#> 96         Wayne  87.168169  6.2273127
#> 97        Wilkes 143.398693 11.2777982
#> 98        Wilson  74.670837  6.1711863
#> 99        Yadkin  95.006796 14.9158944
#> 100       Yancey  62.559665  9.3954516