Download MODIS product files

Need maintenance for the directory path change in NASA EOSDIS. This function first retrieves the all hdf download links on a certain day, then only selects the relevant tiles from the retrieved links. Download is only done at the queried horizontal-vertical tile number combinations. An exception is MOD06_L2 product, which is produced every five minutes every day.

Usage

download_modis_data(
  product = c("MOD09GA", "MOD11A1", "MOD06_L2", "MCD19A2", "MOD13A2", "VNP46A2"),
  version = "61",
  horizontal_tiles = c(7, 13),
  vertical_tiles = c(3, 6),
  mod06_links = NULL,
  nasa_earth_data_token = NULL,
  date_start = "2023-09-01",
  date_end = "2023-09-01",
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)

Arguments

product

character(1). One of c("MOD09GA", "MOD11A1", "MOD06_L2", "MCD19A2", "MOD13A2", "VNP46A2")

version

character(1). Default is "61", meaning v061.

horizontal_tiles

integer(2). Horizontal tile numbers c({start}, {end}). Default is c(7, 13).

vertical_tiles

integer(2). Vertical tile numbers c({start}, {end}). Default is c(3, 6).

mod06_links

character(1). CSV file path to MOD06_L2 download links from NASA LPDAAC. Default is NULL.

nasa_earth_data_token

character(1). Token for downloading data from NASA. Should be set before trying running the function.

date_start

character(1). length of 10. Start date for downloading data. Format YYYY-MM-DD (ex. September 1, 2023 is "2023-09-01").

date_end

character(1). length of 10. End date for downloading data. Format YYYY-MM-DD (ex. September 1, 2023 is "2023-09-01").

directory_to_save

character(1). Directory to save data.

acknowledgement

logical(1). By setting TRUE the user acknowledges that the data downloaded using this function may be very large and use lots of machine storage and memory.

download

logical(1). Download data or only save wget commands.

remove_command

logical(1). Remove (TRUE) or keep (FALSE) the text file containing download commands.

Value

NULL; Raw HDF (.hdf) files will be stored in directory_to_save.

Note

date_start and date_end should be in the same year. Directory structure looks like input/modis/raw/{version}/{product}/{year}/{day_of_year} Please note that date_start and date_end are ignored if product == 'MOD06_L2'.

Author

Mitchell Manware, Insang Song