Download meteorological data (monolevel)

The download_narr_monolevel_data function accesses and downloads monolevel meteorological data from NOAA's North American Regional Reanalysis (NARR) model. "Monolevel" variables contain a single value for the entire atmospheric column (ie. Variable: Convective cloud cover; Level: Entire atmosphere considered as a single layer), or represent a specific altitude associated with the variable (ie. Variable: Air temperature; Level: 2 m).

Usage

download_narr_monolevel_data(
  variables = NULL,
  year_start = 2022,
  year_end = 2022,
  directory_to_save = NULL,
  acknowledgement = FALSE,
  download = FALSE,
  remove_command = FALSE
)

Arguments

variables

character. Variable(s) name acronym. See List of Variables in NARR Files for variable names and acronym codes.

year_start

integer(1). length of 4. Start of year range for downloading data.

year_end

integer(1). length of 4. End of year range for downloading data.

directory_to_save

character(1). Directory(s) to save downloaded data files.

acknowledgement

logical(1). By setting TRUE the user acknowledges that the data downloaded using this function may be very large and use lots of machine storage and memory.

download

logical(1). FALSE will generate a *.txt file containing all download commands. By setting TRUE the function will download all of the requested data files.

remove_command

logical(1). Remove (TRUE) or keep (FALSE) the text file containing download commands.

Value

NULL; Yearly netCDF (.nc) files will be stored in a variable-specific folder within directory_to_save.

Author

Mitchell Manware, Insang Song