Download raw data wrapper function

The download_data() function accesses and downloads atmospheric, meteorological, and environmental data from various open-access data sources.

Usage

download_data(
  dataset_name = c("aqs", "ecoregion", "geos", "gmted", "koppen", "koppengeiger",
    "merra2", "merra", "narr_monolevel", "modis", "narr_p_levels", "nlcd", "noaa",
    "sedac_groads", "sedac_population", "groads", "population", "plevels", "p_levels",
    "monolevel", "hms", "smoke", "tri", "nei", "gridmet", "terraclimate", "huc",
    "cropscape", "cdl", "prism", "olm", "openlandmap"),
  directory_to_save = NULL,
  acknowledgement = FALSE,
  ...
)

Arguments

dataset_name

character(1). Dataset to download.

directory_to_save

character(1). Directory to save / unzip (if zip files are downloaded) data.

acknowledgement

logical(1). By setting TRUE the user acknowledges that the data downloaded using this function may be very large and use lots of machine storage and memory.

...

Arguments passed to each download function.

Note

All download function names are in download_*_data formats

See also

Author

Insang Song